+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# [1.6.0-rc.8](https://github.com/technologiestiftung/wegweiser-frontend/compare/v1.6.0-rc.7...v1.6.0-rc.8) (2023-02-01)
+
+
+### Features
+
+* **WEG-169:** control active value of radio group ([f6f78f4](https://github.com/technologiestiftung/wegweiser-frontend/commit/f6f78f40874e160c5cd6b6db31df6bd826d767a9))
+* **WEG-169:** create first version of radio group component ([1bba8e4](https://github.com/technologiestiftung/wegweiser-frontend/commit/1bba8e4fe084ca484b9bca018930e76d82df0766))
+* **WEG-169:** let radio group change value ([d652bb6](https://github.com/technologiestiftung/wegweiser-frontend/commit/d652bb6919e4c08f0954116824a068d161e05cd4))
+* **WEG-169:** use radio group for focus group selection ([05b2b93](https://github.com/technologiestiftung/wegweiser-frontend/commit/05b2b9338c7dd6d9379b1bb9accbd9b980f6350f))
+
 # [1.6.0-rc.7](https://github.com/technologiestiftung/wegweiser-frontend/compare/v1.6.0-rc.6...v1.6.0-rc.7) (2023-02-01)